refactor: split tome commands across multiple classes for organization
